data/ingest/requirements.txt
langchain-text-splitters
langchain-openai
qdrant-client
pymupdf
python-dotenv